fix tests, add sensor_uid to SensorCalibration
